requirements.txt
torch==2.1.2
torchvision==0.16.2
scikit-learn==1.3.2
Pillow==10.0.0
transformers==4.37.2
datasets==2.19.2
accelerate==0.31.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.13.2
openai==1.35.7
wandb==0.17.2
ftfy==6.2.0
regex==2024.5.15
tqdm==4.66.4
pyarrow==15.0.0